refactor: move partition screen reader table to chart-specific location
