+++ packages/charts/src/chart_types/partition_chart/renderer/canvas/partition.tsx
@@ -17,7 +17,8 @@ import { renderPartitionCanvas2d } from './canvas_renderers';
 import { renderWrappedPartitionCanvas2d } from './canvas_wrapped_renderers';
 import type { Color } from '../../../../common/colors';
 import { Colors } from '../../../../common/colors';
-import { ScreenReaderSummary, ScreenReaderPartitionTable } from '../../../../components/accessibility';
+import { ScreenReaderSummary } from '../../../../components/accessibility';
+import { ScreenReaderPartitionTable } from '../dom/screen_reader_partition_table';
 import { clearCanvas } from '../../../../renderers/canvas';
 import type { SettingsSpec } from '../../../../specs/settings';
 import { onChartRendered } from '../../../../state/actions/chart';
