+++ packages/charts/src/chart_types/partition_chart/renderer/dom/screen_reader_partition_table.tsx
@@ -9,15 +9,15 @@
 import React, { useRef, memo, useState } from 'react';
 import { connect } from 'react-redux';
 
-import type { PartitionData } from '../../chart_types/partition_chart/state/selectors/get_screen_reader_data';
-import { getPartitionScreenReaderDataSelector } from '../../chart_types/partition_chart/state/selectors/get_screen_reader_data';
-import type { SettingsSpec } from '../../specs/settings';
-import type { GlobalChartState } from '../../state/chart_state';
-import type { A11ySettings } from '../../state/selectors/get_accessibility_config';
-import { DEFAULT_A11Y_SETTINGS, getA11ySettingsSelector } from '../../state/selectors/get_accessibility_config';
-import { getInternalIsInitializedSelector, InitStatus } from '../../state/selectors/get_internal_is_intialized';
-import { getSettingsSpecSelector } from '../../state/selectors/get_settings_spec';
-import { isNil } from '../../utils/common';
+import type { PartitionData } from '../../state/selectors/get_screen_reader_data';
+import { getPartitionScreenReaderDataSelector } from '../../state/selectors/get_screen_reader_data';
+import type { SettingsSpec } from '../../../../specs/settings';
+import type { GlobalChartState } from '../../../../state/chart_state';
+import type { A11ySettings } from '../../../../state/selectors/get_accessibility_config';
+import { DEFAULT_A11Y_SETTINGS, getA11ySettingsSelector } from '../../../../state/selectors/get_accessibility_config';
+import { getInternalIsInitializedSelector, InitStatus } from '../../../../state/selectors/get_internal_is_intialized';
+import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_spec';
+import { isNil } from '../../../../utils/common';
 
 interface ScreenReaderPartitionTableProps {
   a11ySettings: A11ySettings;
