+++ packages/charts/src/components/accessibility/index.ts
@@ -8,6 +8,5 @@
 
 /* @internal */
 export { ScreenReaderSummary } from './screen_reader_summary';
-export { ScreenReaderPartitionTable } from './partitions_data_table';
 export { GoalSemanticDescription } from './goal_semantic_description';
 export { AxisDescriptions } from './axis_descriptions';
